+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Assembly instruction can now be dynamic anonymous functions, this is i.e.
   to not convert gifs but force all other images to png.
 
+- Clean unused media feature.
+
 ### Changed
 
 ### Backwards Incompatible Changes